if Xalloca.h is corrected to use size_t param, one change is still needed in libXt to fix compiler warnings - done here.

+++ src/Intrinsic.c
@@ -1388,7 +1388,7 @@ _XtString XtResolvePathname(
     }
 
     bytesAllocd = bytesLeft = 1000;
-    massagedPath = ALLOCATE_LOCAL(bytesAllocd);
+    massagedPath = ALLOCATE_LOCAL((size_t)bytesAllocd);
     if (massagedPath == NULL) _XtAllocError(NULL);
 
     if (path[0] == ':') {
